tugas_1/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
firebase-admin==5.0.0
psutil
python-dotenv